Added GameOfLife project
I am going to start by building the simple game of life in java to help me learn java and to see if I find anything interesting on simulations for a simulation problem I am trying to solve. Later I may recreate this in C# when learning that a bit more for unity if this turns out to be a good learning process.